src/matching_agent.py
#!/usr/bin/env python3
"""
AI Talent Matching Agent

An agentic tool for matching candidates to specific roles using AI.
"""

import argparse
import os
import json
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

def match_candidate_to_role(candidate_profile: dict, job_requirements: dict) -> dict:
    """
    Match a candidate to a job role using AI-powered evaluation.
    
    Args:
        candidate_profile: Dict with candidate skills, experience, etc.
        job_requirements: Dict with job skills, seniority, domain, etc.
    
    Returns:
        Dict with match score and reasoning
    """
    # TODO: Integrate with Claude for intelligent matching
    # - Analyze skill alignment
    # - Evaluate experience level fit
    # - Assess research/domain alignment
    # - Calculate match score (0-100)
    
    match_result = {
        "candidate_id": candidate_profile.get("id"),
        "job_id": job_requirements.get("id"),
        "match_score": 0.0,
        "reasoning": "Matching logic to be implemented",
        "strengths": [],
        "gaps": []
    }
    return match_result

def main():
    parser = argparse.ArgumentParser(description="AI Talent Matching Agent")
    parser.add_argument("--candidate", type=str, help="Path to candidate JSON file")
    parser.add_argument("--job", type=str, help="Path to job requirements JSON file")
    parser.add_argument("--config", type=str, default="config.yaml", help="Configuration file")
    args = parser.parse_args()

    print("AI Talent Matching Agent starting...")
    print(f"Using config: {args.config}")

    if args.candidate and args.job:
        with open(args.candidate, 'r') as f:
            candidate = json.load(f)
        with open(args.job, 'r') as f:
            job = json.load(f)
        
        result = match_candidate_to_role(candidate, job)
        print(f"\nMatch Result:")
        print(json.dumps(result, indent=2))
    else:
        print("Matching agent initialized. Ready for operation.")
        print("Usage: sourcing-agent --candidate <file> --job <file>")

if __name__ == "__main__":
    main()